file-backend/requirements.txt
fastapi[standard]==0.115.12
pydantic==2.11.2
pydantic_core==2.33.1
python-dotenv==1.1.0
redis==5.2.1
SQLAlchemy==2.0.40
PyMySQL==1.1.1
python-jose==3.4.0